Reworked the Timer class. It now uses only the standard library instead of relying on the windows API. Delta time is now in seconds instead of milliseconds.
The system monitor now uses the new timer and a custom clock to get cpu stats. Resulting code looks much cleaner.

+++ ECS Data/src/systems/rendering/camera/camera_system.cpp
@@ -46,8 +46,8 @@ void CameraSystem::processMovement(const Engine& engine, CameraMovement* movemen
 
 	const auto& input = engine.getInput();
 
-	const int2 mouse_delta = input.getMouseDelta();
-	const float dt         = engine.getTimer().deltaTime();
+	const int2  mouse_delta = input.getMouseDelta();
+	const float dt          = static_cast<float>(engine.getTimer().deltaTime());
 
 	float3 rotate_units(0.0f, 0.0f, 0.0f);
 	float3 move_units(0.0f, 0.0f, 0.0f);
